Add support for verifying implementation correctness during starch benchmarks.
User code can provide a STARCH_BENCHMARK_VERIFY function, which is given the
same arguments as the regular function, but always returns bool.

It is called after the implementation to check has been called with the same
arguments. It should verify that the after-call state is correct (e.g. looking
at the contents of output buffers) and return true if all is well, false if
there is a problem.

starch-benchmark will call the verifier for each implementation after setup,
before actually doing the benchmark. If the verifier returns false, that
implementation is skipped with a warning.